requirements.txt
fastapi
uvicorn
python-dotenv
google-genai
langgraph
pydantic